+++ packages/raystack/components/filter-chip/index.tsx
@@ -1,5 +1,6 @@
 export {
   FilterChip,
   type FilterChipCalendarProps,
+  type FilterChipProps,
   type FilterChipValue
 } from './filter-chip';